Attempt to create wgpu surface in winit
I've fallen for one of the classic traps, which is to try and call an
async function from inside a sync function. After reading some posts on
Stack Overflow, GitHub, and the documentation, I think I don't have the
appetite for solving this. I'll leave you with this quote from a winit
issue 1199:

> One fairly significant papercut I've run into when working on my own
> async event loop wrapper is that it's very difficult to manage the
> lifetimes required to create a window during the execution of the
> Future.