Generate custom GLAD loading library
This moves GLAD from vcpkg to CMake using FetchContent. This allows us
to run GLAD to generate bindings for the API of our choice. For now the
API is OpenGL 3.3 core profile, but in the future we may want KHR_debug
or other specific extensions, and may want to choose a different
version.

This adds Python and Jinja2 as dependencies on Windows.