Generate all OpenGL entry points in feature set
Functions which are excluded are now replaced with a stub that calls
MissingFunction(), which aborts the program. This may improve the code
sense experience when developing new code, since you can get working
code completion for functions that are excluded (and they will
automatically be included on the next rebuild).